Move test_context to helper library
